Consolidate Container Hook bootstrap logic into Overrider (#132)
Removes container hook from executor, in favor of using Overrider as
the unified bootstrap process.

Container hook has two tasks when STATSD_UDP_HOST/_PORT are set:
- Copy 2 libs into kafka's libs dir
- Enable metrics output in kafka's config

With this commit:
- The jars are included in overrider.zip:/metrics-libs
- The jars are copied to the kafka dir configured for the framework.
  Container hook meanwhile had a DIY search for "kafka_*" which
  ignored the config.
- Then the extra metrics config vals are added to the kafka props file

While the container hook is no longer referenced by the framework as
of this commit, the code is still left in the repo to avoid breaking
CI which still builds the hook package. The hook code will be deleted
in a separate commit (in parallel with a CI update).