Enable datadog tags in Kafka's statsd output to the mesos module (#165)
This was only disabled for better compatibility with users of the old
'metrics.marathon.mesos' forwarding method. Under the new method, the
statsd tags are automatically converted to tags in the avro schema.
Consumers can then decide what to do with the resulting tag data.
Including the tags allows for richer output in dashboards.